refactor(db): extract 7 more entity modules — decomposition Wave 2
Wave 2: concepts, projects(+orgs), operators(+studio users), approvals(+GATED_ACTIONS),
runs, messages(+inbox/requests/team-chat), context. Bodies verbatim; barrel re-exports
via `export *`.

- channels DEFERRED: its blueprint imports listAgents from ./agents.js (not yet
  extracted); acyclic extraction needs agents first (later wave). Channel functions stay
  resident for now — no behavior change.
- export* scope trap: 21 resident call sites (boot-payload/overview/health) called moved
  helpers by bare name -> explicit imports added to db.js (acyclic).
- context: master inlines prototype/data sanitization (no helper fns as in the blueprint)
  -> moved master's actual code verbatim.

db.js: 3995 -> 3159 (-836). Gate: manifest --check green (308); suite 45 files / 310.

Executor: GLM-5.2 (agentic, gate-driven). Reviewed + verified by m5Max.
